more glossary updates
