+++ frontend/cfb/resources/glossary.js
@@ -1,7 +1,7 @@
 const logger = require("../../utils/logger");
 const fs = require("fs/promises")
 const path = require("path")
-const { alphabetize } = require("../../utils/misc");
+const { ALPHABET } = require("../../utils/misc");
 
 async function generateGlossaryItems() {
     const glossaryPath = path.resolve(__dirname, "..", "..", "static", "glossary.json");
@@ -9,7 +9,17 @@ async function generateGlossaryItems() {
         logger.info(`Loading glossary from path ${glossaryPath}...`)
         const data = await fs.readFile(glossaryPath, { encoding: "utf-8" })
         const tmpGloss = JSON.parse(data);
-        const glossary = alphabetize(tmpGloss)
+        let glossary = {};
+        ALPHABET.forEach(letter => {
+            let records = tmpGloss[letter];
+            if (records) {
+                let copyRec = [...records];
+                copyRec.sort((a, b) => {
+                    return a.term.localeCompare(b.term)
+                });
+                result[letter] = copyRec;
+            }
+        });
 
         logger.info(`Loaded glossary for ${Object.keys(glossary)} letters`)
         return glossary;
